@@ -1,19 +1,17 @@
 import agent from "../agent";
 import Header from "./Header";
-import React from "react";
+import React, { useEffect } from "react";
 import { connect } from "react-redux";
 import { APP_LOAD, REDIRECT } from "../constants/actionTypes";
-import { Route, Switch } from "react-router-dom";
-import Item from "../components/Item";
-import Editor from "../components/Editor";
-import Home from "../components/Home";
-import Login from "../components/Login";
-import Profile from "../components/Profile";
-import ProfileFavorites from "../components/ProfileFavorites";
-import Register from "../components/Register";
-import Settings from "../components/Settings";
-import { store } from "../store";
-import { push } from "react-router-redux";
+import Item from "./Item";
+import Editor from "./Editor";
+import Home from "./Home";
+import Login from "./Login";
+import Profile from "./Profile";
+import ProfileFavorites from "./ProfileFavorites";
+import Register from "./Register";
+import Settings from "./Settings";
+import { Route, Routes, useNavigate } from "react-router-dom";
 
 const mapStateToProps = (state) => {
   return {
@@ -30,59 +28,54 @@ const mapDispatchToProps = (dispatch) => ({
   onRedirect: () => dispatch({ type: REDIRECT }),
 });
 
-class App extends React.Component {
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.redirectTo) {
-      // this.context.router.replace(nextProps.redirectTo);
-      store.dispatch(push(nextProps.redirectTo));
-      this.props.onRedirect();
-    }
-  }
+const App = (props) => {
+  const { redirectTo, onRedirect, onLoad } = props;
+  const navigate = useNavigate();
 
-  componentWillMount() {
+  useEffect(() => {
+    if (redirectTo) {
+      navigate(redirectTo);
+      onRedirect();
+    }
+  }, [redirectTo, onRedirect, navigate]);
+
+  useEffect(() => {
     const token = window.localStorage.getItem("jwt");
     if (token) {
       agent.setToken(token);
     }
+    onLoad(token ? agent.Auth.current() : null, token);
+  }, [onLoad]);
 
-    this.props.onLoad(token ? agent.Auth.current() : null, token);
-  }
-
-  render() {
-    if (this.props.appLoaded) {
-      return (
-        <div>
-          <Header
-            appName={this.props.appName}
-            currentUser={this.props.currentUser}
-          />
-          <Switch>
-            <Route exact path="/" component={Home} />
-            <Route path="/login" component={Login} />
-            <Route path="/register" component={Register} />
-            <Route path="/editor/:slug" component={Editor} />
-            <Route path="/editor" component={Editor} />
-            <Route path="/item/:id" component={Item} />
-            <Route path="/settings" component={Settings} />
-            <Route path="/@:username/favorites" component={ProfileFavorites} />
-            <Route path="/@:username" component={Profile} />
-          </Switch>
-        </div>
-      );
-    }
+  if (props.appLoaded) {
     return (
       <div>
         <Header
-          appName={this.props.appName}
-          currentUser={this.props.currentUser}
+          appName={props.appName}
+          currentUser={props.currentUser}
         />
+        <Routes>
+          <Route exact path="/" element={<Home/>} />
+          <Route path="/login" element={<Login/>} />
+          <Route path="/register" element={<Register/>} />
+          <Route path="/editor/:slug" element={<Editor/>} />
+          <Route path="/editor" element={<Editor/>} />
+          <Route path="/item/:id" element={<Item/>} />
+          <Route path="/settings" element={<Settings/>} />
+          <Route path="/:username/favorites" element={<ProfileFavorites/>} />
+          <Route path="/:username" element={<Profile/>} />
+        </Routes>
       </div>
     );
   }
+  return (
+    <div>
+      <Header
+        appName={props.appName}
+        currentUser={props.currentUser}
+      />
+    </div>
+  );
 }
-
-// App.contextTypes = {
-//   router: PropTypes.object.isRequired
-// };
 
 export default connect(mapStateToProps, mapDispatchToProps)(App);
