@@ -9,6 +9,7 @@ import {
   PROFILE_PAGE_LOADED,
   PROFILE_PAGE_UNLOADED,
 } from "../constants/actionTypes";
+import { withRouterParams } from "./commons";
 
 const EditProfileSettings = (props) => {
   if (props.isUser) {
@@ -76,11 +77,12 @@ const mapDispatchToProps = (dispatch) => ({
 });
 
 class Profile extends React.Component {
-  componentWillMount() {
+  componentDidMount() {
+    const username = this.props.params.username?.substring(1);
     this.props.onLoad(
       Promise.all([
-        agent.Profile.get(this.props.match.params.username),
-        agent.Items.bySeller(this.props.match.params.username),
+        agent.Profile.get(username),
+        agent.Items.bySeller(username),
       ])
     );
   }
@@ -166,5 +168,5 @@ class Profile extends React.Component {
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Profile);
+export default connect(mapStateToProps, mapDispatchToProps)(withRouterParams(Profile));
 export { Profile, mapStateToProps };
