@@ -7,6 +7,7 @@ import {
   PROFILE_PAGE_LOADED,
   PROFILE_PAGE_UNLOADED,
 } from "../constants/actionTypes";
+import { withRouterParams } from "./commons";
 
 const mapDispatchToProps = (dispatch) => ({
   onLoad: (pager, payload) =>
@@ -15,12 +16,13 @@ const mapDispatchToProps = (dispatch) => ({
 });
 
 class ProfileFavorites extends Profile {
-  componentWillMount() {
+  componentDidMount() {
+    const username = this.props.params.username?.substring(1);
     this.props.onLoad(
-      (page) => agent.Items.favoritedBy(this.props.match.params.username, page),
+      (page) => agent.Items.favoritedBy(username, page),
       Promise.all([
-        agent.Profile.get(this.props.match.params.username),
-        agent.Items.favoritedBy(this.props.match.params.username),
+        agent.Profile.get(username),
+        agent.Items.favoritedBy(username),
       ])
     );
   }
@@ -33,7 +35,7 @@ class ProfileFavorites extends Profile {
     return (
       <ul className="nav nav-tabs outline-active">
         <li className="nav-item">
-          <Link className="nav-link" to={`/@${this.props.profile.username}`}>
+          <Link className="nav-link" to={`/@${this.props.profile?.username}`}>
             My Items
           </Link>
         </li>
@@ -41,7 +43,7 @@ class ProfileFavorites extends Profile {
         <li className="nav-item">
           <Link
             className="nav-link active"
-            to={`/@${this.props.profile.username}/favorites`}
+            to={`/@${this.props.profile?.username}/favorites`}
           >
             Favorited Items
           </Link>
@@ -51,4 +53,4 @@ class ProfileFavorites extends Profile {
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(ProfileFavorites);
+export default connect(mapStateToProps, mapDispatchToProps)(withRouterParams(ProfileFavorites));
