@@ -2,8 +2,8 @@ const axiosLib = require("axios");
 const fs = require("fs");
 
 const WILCO_ID = process.env.WILCO_ID || fs.readFileSync('../.wilco', 'utf8')
+const baseURL = process.env.ENGINE_BASE_URL || "https://engine.wilco.gg"
 
-const baseURL = 'https://wilco-engine.herokuapp.com';
 const axios = axiosLib.create({
   baseURL: baseURL,
   headers: {
@@ -12,8 +12,12 @@ const axios = axiosLib.create({
 });
 
 async function sendEvent(event, metadata) {
-  const result = await axios.post(`/users/${WILCO_ID}/event`, JSON.stringify({ event, metadata }));
-  return result.data;
+    try {
+        const result = await axios.post(`/users/${WILCO_ID}/event`, JSON.stringify({event, metadata}));
+        return result.data;
+    } catch (error) {
+        console.error(`failed to send event ${event} to Wilco engine`)
+    }
 }
 
 module.exports = {
