@@ -2,20 +2,16 @@ import "./custom.scss";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import React from "react";
-import { store, history } from "./store";
-
-import { Route, Switch } from "react-router-dom";
-import { ConnectedRouter } from "react-router-redux";
+import { store } from "./store";
 
 import App from "./components/App";
+import { BrowserRouter } from "react-router-dom";
 
 ReactDOM.render(
   <Provider store={store}>
-    <ConnectedRouter history={history}>
-      <Switch>
-        <Route path="/" component={App} />
-      </Switch>
-    </ConnectedRouter>
+    <BrowserRouter basename="/">
+      <App />
+    </BrowserRouter>
   </Provider>,
 
   document.getElementById("root")
